backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Security scanning
bandit==1.7.6
safety==3.0.1

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Rate limiting
slowapi==0.1.9

# Redis for caching
redis==5.0.1
hiredis==2.3.2

# JWT
pyjwt==2.8.0

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Environment
environs==10.3.0
